swap, amd64 - increase maximum swap space to 1TB x 4
* The radix can overflow a 32 bit integer even if swblk_t fits in 32 bits.
  Expand the radix to 64 bits and thus allow the subr_blist code to operate
  up to 2 billion blocks (8TB total).

* Shortcut the common single-swap-device case.  We do not have to scan
  the radix tree to get available space in the single-device case.

* Change maxswzone and maxbcache to longs and add TUNABLE_LONG_FETCH().

* All the TUNEABLE_*_FETCH() calls and kgetenv_*() calls for integers
  call kgetenv_quad().

  Adjust kgetenv_quad() to accept a suffix for kilobytes, megabytes,
  gigabytes, and terrabytes.

+++ sys/sys/buf.h
@@ -365,8 +365,8 @@ struct cluster_save {
 
 #ifdef _KERNEL
 extern int	nbuf;			/* The number of buffer headers */
-extern int	maxswzone;		/* Max KVA for swap structures */
-extern int	maxbcache;		/* Max KVA for buffer cache */
+extern long	maxswzone;		/* Max KVA for swap structures */
+extern long	maxbcache;		/* Max KVA for buffer cache */
 extern int	runningbufspace;
 extern int	runningbufcount;
 extern int	hidirtybufspace;
